Remove cookiecutter; bump infra beman-submodule; update CMake style for pre-commit
Cookiecutter is intended to help users create new libraries, but isn't
needed outside of exemplar; this commit removes it.

infra/cmake/use-fetch-content.cmake became out of sync with the
upstream due to a bad find-and-replace; this commit brings it back
into alignment with infra and updates the other files in infra to the
latest version.

Finally, pre-commit was failing because some CMake formatting
disagreed with the gersemi style checker used by pre-commit; this
commit changes applies that tool's changes.