Updated list of supported implementations
Changed status to "Under development and not yet ready for production
use".

Updated Compiler Explorer example to not use std::jthread since that is
missing in libc++ and to not use chrono operator<< overloads to display
durations since some of the older implementations that this library
support have not implemented those overloads. Made the same adjustments
to the test and example code.

Updated the CI workflows so they don't try to use ThreadSanitizer when
using libstdc++ versions that won't compile when using std::timed_mutex.
Also removed compiling in C++17 mode for now. It should be be easy to
replace the concepts with old-school type traits if needed.

Signed-off-by: Ted Lyngmo <ted@lyncon.se>